➕ Code cleaning for mysql, postgres and async
* Moved common functions to a common package
* Created common Record for the DB supported
